my_ETL/requirements.txt
pymongo
sqlalchemy
vaderSentiment
pandas
psycopg2-binary